requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
pip==24.0
sqlparse==0.5.0
